move tests, to allow esbuild direct
