+++ oceanics-io-api/test/index.spec.ts
@@ -1,5 +1,5 @@
 import { describe, expect, test } from '@jest/globals';
-import { EXTENSIONS, apiFetch, testAllowedMethodCount } from "./shared/test-utils";
+import { EXTENSIONS, apiFetch, testAllowedMethodCount } from "./test-utils";
 
 describe("idempotent", function () {
   /**
